reorganizes files to have better structure and clear definition
